backend/requirements.txt
# For API
fastapi==0.68.0
uvicorn==0.15.0
python-multipart

# For yoinking images
Pillow

# For model
scikit-learn
scikeras
keras
tensorflow